Activate "header by height" support via a config option
Currently only BCH Unlimited support a block height as a parameter
to the getblockheader RPC call. In commit f69281c0d we assume that
all BCH clients support this which is not the case.

Since what we aim for this explorer is to work with all BCH clients
implementations, a new configuration parameter has been added to turn
this "header by height" functionality on/off (BTCEXP_HEADER_BY_HEIGHT_SUPPORT).