Fix 404/500 properly
Due to the way perf measurements where implemented, ie hoping to a
new app.use() call in the routing, 404/500 and error handling in general
were borked. To fix that a new utils funciton has been implememted and
use everytime we need to measure the perf of a given route. That
way we could remove a bunch of calls to next() (maybe to much but
I guess we will discover soon)