Fix Redis cache with unknown miners
The `getMinerFromCoinbaseTx` would return an array instead of an
object, which caused Redis caching to fail. The Redis cache relies
on the block object passing through `JSON.stringify` / `JSON.parse`,
which can not handle named keys on an array.

This also reduces code duplication by breaking through the outer
loop and determining `coinbaseAscii` and `noIFP` once at the end
of the function.