Fix miner info overwrite caused modifying the same object
The `getMinerFromCoinbaseTx` takes references to the data in
`miningPoolsConfig`, modifies them and returns them. This causes
the code to show the last modification when rendering templates,
but put the correct values into the cache.

Instead of modifying the existing object, we use `Object.assign`
to copy the values from `miningPoolsConfig` into our return object.