API: cache `getblock` requests with partial information
First, this unifies all the `getBlock` and `getBlockHeader` functions to
one common function in the RPC API. Handling height stuff is done in
Core API.

The Core API adds a new parameter `full`, when it is set to `false`, the
block will not be cached, when it is set to `true`, the block will be
cached with one year expiry, but only containing the coinbase
transaction. Also the miner info will not be cached to allow quickly
updating the miner configs without needing to flush the cache.

The calls getting multiple blocks were simplified to use JS `.map()`.

All get `getBlock` calls for the search were changed to use
`getBlockHash` instead, which saves a lot of time with big blocks.

The `/block-height` and `/block-analysis` page were changed to not
fetch the same block twice.

Call sites where the TX data is not required were changed to not fetch
the full block, benefitting from caching.

The block list relied on the TX data to show TX counts, this has been
replaced with the blocks `nTx` counter. Not all nodes support the `nTx`
property, so if it is undefined we compute it from `tx.length`.

Overall this give a significant speedup to many parts of the explorer.