Redis: Use MessagePack for a ~30% reduction in memory usage
`JSON` is not the best suited format to store in Redis given its large
overhead. Redis supports `MessagePack` in its scripting language, so it
can be used as a drop in replacement. A disadvantage is that the cached
keys are not human readable anymore, but the JSON was hardly any better.

This adds a new dependency on the NPM `msgpack-lite` library. There are
faster ones available but the performance is quite good already.

Note that for further reductions in memory usage a fast compression
algorithm like `zstd` could be used.