Use dynamic chunk size when loading mempool
This uses a larger chunk size when loading a lot of transactions. This
could probably be better optimized for example by passing a index into
a mempool snapshot, or just sending all TX, which is feasible. Chunk
size can't be increased much further because of HTTP limitations,
unless it is switched to HTTP PUSH.