Speed up block detail page, only fetch TX when needed
Previously the block detail page would send *every* TX id in the block
as a giant syntax highlighted JSON array, which would significantly
increase the time required to build the response, the time to transfer
the response, the time to render the site as well as the memory usage of
the site. On 256MB blocks, the site would take ~1.4GB of memory.

This removes this list and instead AJAX loads it as plain JSON from an
API endpoint. It uses highlight.js in the browser to do the syntax
highlighting on the client side. With this change, big blocks load
really fast and users can still get the TX list if they need it.