Add better unified coinbase display with potential signals
Moves the code from `block-content.pug` and `blocks-list.pug` into an
includable file that needs to have the `block` variable set as an input.

Unifies how the coinbase string is displayed in the blocks list and
block details page.

This also displays suspected signals in a 3rd color. A suspected signal
is any string between two / characters with greedy matching.