Add .editorconfig file
There seems to be an issue where trailing whitespace disappears in .pug
template files. At least with editors supporting it, this can be
prevented by including an .editorconfig file. The config is set to use
LF for line endings, tab for indentation, use UTF-8 for JavaScript
source and preserve trailing whitespace in pug files.